agendadj/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-model-utils==4.1.1
djangorestframework==3.12.4
pkg==0.2
psycopg2-binary==2.8.6
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1
